+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["cel-parser", "ic-response-verification-rs", "representation-independent-hashing", "examples/rust"]
+members = ["packages/ic-response-verification", "examples/rust"]
