+++ packages/ic-response-verification/src/cel.rs
@@ -1,0 +1,12 @@
+mod ast_mapping;
+mod error;
+mod parser;
+
+use crate::cel::ast_mapping::map_cel_ast;
+use crate::cel::error::CelParserResult;
+use crate::cel::parser::parse_cel_expression;
+use crate::types::Certification;
+
+pub fn cel_to_certification(cel: &str) -> CelParserResult<Option<Certification>> {
+    parse_cel_expression(cel).and_then(map_cel_ast)
+}
