--- representation-independent-hashing/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "representation-independent-hashing"
-version = "0.0.0"
-edition = "2021"
-
-[dependencies]
-http = "0.2.8"
-leb128 = "0.2.5"
-sha2 = "0.10.6"
-cel-parser = { path = "../cel-parser" }
-
-[dev-dependencies]
-hex = "0.4.3"
