--- representation-independent-hashing/src/lib.rs
@@ -1,5 +1,0 @@
-mod hash;
-mod representation_independent_hash;
-
-pub mod request_hash;
-pub mod response_hash;
